Always send the version header
Starting with version 4 the oVirt engine will support multiple versions
of the API, initially versions 3 and 4. The caller will be able to
select the version using the "Version" header. Versions 3 and 4 of the
API aren't compatible, and this gem only works with version 3. This
patch changes the the gem so that it always explicitly requests version
3 of the API, so it will work with older versions of the engine (which
will just ignore the header) as well as with the new version.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>