Probe multiple API paths
Currently this gem assumes that the path of the API is "/api", but in fact it
may be "/api" or "/ovirt-engine/api" since version 3.5 of the engine, and only
"/ovirt-engine/api" in version 4 of the engine. This means that currently this
gem can't work with version 4 of the engine, because it will always use an
incorrect path. To avoid that, while preserving compatibility with earlier
versions of the engine, this patch changes the gem so that it probes the paths.
It will then select the first path where the probe succeeds.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>