Configurable API prefix
Currently this gem assumes that the oVirt API is always located in the
"/api" path, but in recent versions of oVirt (since version 3.5) it is
located both in "/api" and "/ovirt-engine/api". Starting with version 4
of the engine the "/api" path will be removed. To handle this change in
advance this patch modifies the gem so that it can receive explicitly
from the caller the path to use.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>