Use version 2.0.0.rc1 of rest-client
This is the version used by ManageIQ itself, and is good to be in sync.
In addition this version of rest-client changed the signature of the
Request#return! method, so we need to change the way we call it,
otherwise unexpected response codes returned by the engine result in
exceptions and failed inventory refreshes, for example.

Bug-Url: https://bugzilla.redhat.com/1328087
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>