Remove "ovirt-engine" prefix from paths
The path passed to the "api_uril" method will have the "/api" prefix if it is
extracted from the "ems_ref" attribute stored in the database, and will have
the "/ovirt-engine/api" if it comes directly from the "href" attribute of the
XML documents, for example when using the "relationships" method to fetch
secondary objects related to the primary object. This means that to have a
clean path we need to remove both "ovirt-engine" and "api".

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>