Update Guaranteed memory when updating VM's memory
There is a coupling between the Guaranteed memory to the VM memory:
Guaranteed cannot exceed the VM memory, and the VM memory cannot
be set to a lower value than the guaranteed memory.

The ovirt-engine UI behavior sets the same value for both memory
and guaranteed - this patch imitates the same behavior.

http://bugzilla.redhat.com/1356193