Using collection fetching during targeted refresh
When we have not enough information how to fetch specific storage
information we need to get all storage domains for specific
cluster and later match it with other information fetched during
refresh.

In order to enable above scenario we need to be able to fetch
collections when running targeted refresh.