Fix discovery of version 4 of oVirt
Currently this gem tests for the presence of the oVirt engine sending a
request to the following URL:

  http://.../engine.ssh.key.txt

This doesn't work with version 4 of the engine, as that path was
removed. In order to support older and newer versions of the engine this
patch modifies the gem so that it checks first the path that works since
version 3.5 of the engine, and then, if that doesn't work, the old path.

Bug-Url: https://bugzilla.redhat.com/1382732
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>