Add support for matrix parameters to the 'update' method
The oVirt API supports matrix parameters that alter the meaning of some
operations, including updates. For example, the operation to update a
virtual machine can receive an optional 'next_run' parameter that
controls if the update is applied to the state of the current running
virtual machine (the default) or only to the next execution of the
virtual machine, after rebooting it. For example, to update the memory
of for the next boot of the virtual machine the following request should
be used:

  PUT /ovirt-engine/api/vms/123;next_run=true

This mechanism isn't currently supported by this gem. That patch adds
support for it.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>