Add 'matrix' parameters to methods to update memory
Currently the methods 'memory=' and 'memory_reserve=' of the 'Vm' class
don't support adding matrix parameters, and thus they don't support
adding the 'next_run' parameter which is very relevant when updating
virtual machines. This patch adds a ne method named 'update_memory'
that can update simultaneously the virtual and guaranteed memory, and
that supports matrix parameters. The 'memory=' and 'memory_reserve='
methods are then re-implemented as simple calls to the new method.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>