Fix handling of version 4.y in Vm.attach_payload
The Vm.attach_payload method has two versions: one for oVirt older than
3.3, and another for oVirt 3.3 or newer. But the 'payloads_version' that
decides which one to use does not take into account that the major
version can be greater than 3. As a result when the major version is 4.0
it selects the version of the method for oVirt older than 3.3. This
patch fixes that issue.

https://bugzilla.redhat.com/1441746