Fix host targeted refresh
In V3 the host resources retrieved by
/ovirt-engine/api/hosts/{host:id} contains the statistics element.
The statistics element contains a reference to the host element.
The xpath used to collect similar host elements considered the
host sub-element of the statistics as a resource to collect,
which lead to collecting the same host multiple times.
By specifying the exact top level host path, this will be avoided.

https://bugzilla.redhat.com/show_bug.cgi?id=1448690