Create trusted CA certificates file when needed
Currently the gem creates the temporary file contining the trusted CA
certificates when the 'Service' class is instantiated, and removes it
when the 'disconnect' method is called. But there are situations where
the instance is used *after* the 'disconnect' method has been called. In
those situations the trusted CA certificates won't be available, and the
connection will fail. To avoid that problem this patch changes the
'Service' class so that the temporary file is created when needed,
regardless of when the 'disconnect' method is called.

This patch addresses the following bug:

  Service order request for VM provision from template fail on SSL Certificate verification
  https://bugzilla.redhat.com/1483303